pdf-sempart/pyproject.toml
[project]
name = "pdf-sempart"
version = "0.1.0"
description = "Layout-aware semantic chunking for PDFs using hLDA and dual embeddings."
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "sentence-transformers>=3.0.0",
  "tomotopy>=0.13.0",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy", "types-PyYAML"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
